rename request-config.ts to request.ts
